feat: users + products + productCart completed with DB CRUD

+++ Project/database/models/Product.js
@@ -21,7 +21,7 @@ module.exports = function (sequelize, DataTypes) {
       allowNull: false
     },
     price: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(9,2),
       allowNull: false
     },
     image: {
